requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
bs4==0.0.1
Django==4.0.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
lxml==4.7.1
pytz==2021.3
soupsieve==2.3.1
sqlparse==0.4.2
tzdata==2021.5
